12/4

+++ kiosk/src/view/DeleteProductView.java
@@ -11,17 +11,28 @@ public class DeleteProductView {
 		Scanner sc = new Scanner(System.in);
 		ProductDAO pdao = new ProductDAO();
 		
-		System.out.println("------ 메뉴 변경 ------");
-		
-		//내가 올린 상품 목록 띄우기
-//		System.out.println(pdao.getList(loginUser.userid));
-		System.out.println("삭제할 메뉴 이름 : ");
-		String menu_name = sc.next();
-		
-		if(pdao.deleteProduct(cus_id, menu_name)) {
-			System.out.println(menu_name+"이 정상적으로 삭제되었습니다.");
-		}else {
-			System.out.println("메뉴 삭제 실패 / 다음에 다시 입력해주세요.");
+		while(true) {
+			System.out.println("------ 메뉴 변경 ------");
+			System.out.println("[1] 메뉴 변경");
+			System.out.println("[2] 장바구니로 다시 돌아가기");
+			int number = sc.nextInt();
+			
+			if(number == 1) {
+				//내가 올린 상품 목록 띄우기
+//				System.out.println(pdao.getList(cus_id));
+				System.out.println("삭제할 메뉴 이름 : ");
+				String menu_name = sc.next();
+				
+				if(pdao.deleteProduct(cus_id, menu_name)) {
+					System.out.println(menu_name+"이 정상적으로 삭제되었습니다.");
+				}else {
+					System.out.println("메뉴 삭제 실패 / 다음에 다시 입력해주세요.");
+				}
+			}else if(number == 2) {
+				new PaymentView();
+			}else {
+				System.out.println("번호를 다시 입력해주세요.");
+			}
 		}
 	}
 }
